Fix onUse not firing when right-clicking air. Closes GH-1
If target was undefined, the 'use' event handler would return
immediately. Reworked the conditionals so that the onUse event handler
is called regardless. Also, prioritized onUse over itemblock placement;
if there is an onUse property for a block, it will be used instead of
placing the block.

Allows you to, for example, eat with voxel-food while you are not
looking at a block.