Fix no target check in useBlock
Previously, if a block was used (right-clicked) with no raycasted voxel
(target === false), @game.canCreateBlock would be called with undefined.
This was expected to always return false, and it did with the
three.js-based voxel-engine, but not with the ndarray-based voxel-engine
branch. Fix this discrepancy relying on undefined behavior by adding a
check for the target in useBlock, returning immediately if falsey.